Initiatives: extract ThresholdReachedMessage
