+++ front/app/containers/Admin/initiatives/settings/VotingLimit.tsx
@@ -5,23 +5,18 @@ import { SectionField, SubSectionTitle } from 'components/admin/Section';
 import { Input } from 'cl2-component-library';
 import Error from 'components/UI/Error';
 import errorMessages from 'components/UI/Error/messages';
-import { StyledWarning, FormValues } from '.';
+import { StyledWarning, FieldProps } from '.';
 
 // i18n
 import { injectIntl, FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
 import { InjectedIntlProps } from 'react-intl';
 
-interface Props {
-  formValues: FormValues;
-  setParentState: (state: any) => void;
-}
-
 const VotingLimit = ({
   formValues,
   setParentState,
   intl: { formatMessage },
-}: Props & InjectedIntlProps) => {
+}: FieldProps & InjectedIntlProps) => {
   const handleDaysLimitOnChange = (value: string) => {
     setParentState(({ formValues }) => ({
       formValues: {
